requirements.txt
# CareLens 360 - Dependencies

# Streamlit for UI
streamlit>=1.28.0

# Google Cloud Services
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0
google-generativeai>=0.3.0

# Image Processing
Pillow>=10.0.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2

